refactor: update data retention, maintenance, and module toggle config schemas, removing unused interfaces and adding createdBy field
