+++ packages/data/schema/system/src/model/system-settings/system_behave/maintenance_config.schema.ts
@@ -1,6 +1,18 @@
 //@ts-ignore
 import mongoose, { Schema } from 'mongoose';
-import { IMaintenanceConfig } from './interfaces';
+//import { IMaintenanceConfig } from './interfaces';
+import { ICreatedByAdmin } from '@isyss-cdm/interface';
+
+
+export interface IMaintenanceConfig extends Document {
+	id?: string;
+	maintenanceMode?: Record<string, unknown>; // could be boolean or object describing mode
+	affectedModules?: Array<Record<string, unknown>>; // jsonb list
+	createdAt?: Date;
+	updatedAt?: Date;
+	createdBy?: ICreatedByAdmin;
+}
+
 
 const baseFields = {
 	id: { type: String, index: true },
@@ -26,5 +38,5 @@ export const MaintenanceConfigModel =
 	(mongoose.models.MaintenanceConfig as mongoose.Model<IMaintenanceConfig>) ||
 	mongoose.model<IMaintenanceConfig>('MaintenanceConfig', MaintenanceConfigSchema);
 
-export { IMaintenanceConfig };
+//export { IMaintenanceConfig };
 export default MaintenanceConfigModel;
